kloud/rackspacke: check if image is existing
