+++ go/src/koding/kites/kloud/api/openstack/image.go
@@ -1,0 +1,29 @@
+package api
+
+import (
+	"fmt"
+
+	"github.com/rackspace/gophercloud"
+)
+
+type Images []gophercloud.Image
+
+func (i Images) String() string {
+	out := ""
+	for d, image := range i {
+		out += fmt.Sprintf("[%d] name: %+v id: %+v\n", d, image.Name, image.Id)
+	}
+
+	return out
+}
+
+// Image returns a single image based on the given image id, slug or id. It
+// checks for each occurency and returns for the first match.
+func (o *Openstack) Image(id string) (*gophercloud.Image, error) {
+	return o.Client.ImageById(id)
+}
+
+// Images returns all available images for this account
+func (o *Openstack) Images() (Images, error) {
+	return o.Client.ListImages()
+}
